move signer to separate class
